moved ping to network.go
